refactor: 템플릿 메서드를 다시 전략패턴으로 수정 사유: 중복로직 많이 없음, default로 해결가능
