+++ src/main/java/com/coop/global/notification/core/NotificationProducer.java
@@ -2,10 +2,9 @@ package com.coop.global.notification.core;
 
 import com.coop.global.common.enums.ErrorCode;
 import com.coop.global.exception.error.InvalidRequestException;
-import com.coop.global.notification.templates.NotificationTemplate;
+import com.coop.global.notification.templates.NotificationStrategy;
 import com.coop.global.notification.values.TriggerNotification;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.Around;
 import org.aspectj.lang.annotation.Aspect;
@@ -16,22 +15,21 @@ import java.util.List;
 @Aspect
 @Component
 @RequiredArgsConstructor
-@Slf4j
 public class NotificationProducer {
 
-    private final List<NotificationTemplate> templates;
+    private final List<NotificationStrategy> strategies;
 
     @Around("@annotation(triggerNotification)")
     public Object notify(ProceedingJoinPoint joinPoint, TriggerNotification triggerNotification) throws Throwable {
         Object result = joinPoint.proceed();
         Object[] args = joinPoint.getArgs();
 
-        NotificationTemplate template = templates.stream()
-                .filter(t -> t.supports(triggerNotification.target()))
+        NotificationStrategy strategy = strategies.stream()
+                .filter(s -> s.supports(triggerNotification.target()))
                 .findFirst()
                 .orElseThrow(() -> new InvalidRequestException(ErrorCode.INVALID_NOTIFICATION_TYPE));
 
-        template.execute(args, result, triggerNotification.target());
+        strategy.execute(args, result);
         return result;
     }
 }
